+++ src/OTPublisher.js
@@ -155,6 +155,11 @@ class OTPublisher extends Component {
   getRtcStatsReport() {
     OT.getRtcStatsReport(this.state.publisherId);
   }
+
+  setVideoTransformers(videoTransformers) {
+    OT.setVideoTransformers(this.state.publisherId, videoTransformers);
+  }
+
   render() {
     const { publisher, publisherId } = this.state;
     const { sessionId } = this.context;
@@ -176,6 +181,7 @@ OTPublisher.propTypes = {
   properties: PropTypes.object, // eslint-disable-line react/forbid-prop-types
   eventHandlers: PropTypes.object, // eslint-disable-line react/forbid-prop-types
   getRtcStatsReport: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+  setVideoTransformers: PropTypes.func, // eslint-disable-line react/forbid-prop-types
 };
 OTPublisher.defaultProps = {
   properties: {},