+++ CHANGELOG.md
@@ -11,6 +11,10 @@
   [known issues](https://tokbox.com/developer/sdks/ios/release-notes.html#known-issues)
   in the OpenTok iOS SDK release notes.
 
+- [Update]: The new OTPublisher.setVideoFilters() method lets you set (and clear)
+  video transformers for a publisher. For more info, see the docs:
+  [OTPublisher](/docs/OTPublisher.md).
+
 - [Fix]: Fixes an issue in which applications could not connect to a session when
   the `proxyUrl` option for OTSession was set. - issue #645